fix: rename util file to event.js and move loadData to it
