+++ src/components/map/layers/EventLayer.js
@@ -2,8 +2,8 @@ import { getInstance as getD2 } from 'd2'
 import React from 'react'
 import { EVENT_COLOR, EVENT_RADIUS } from '../../../constants/layers.js'
 import { getContrastColor } from '../../../util/colors.js'
+import { getAnalyticsRequest } from '../../../util/event.js'
 import { filterData } from '../../../util/filter.js'
-import { getAnalyticsRequest } from '../../../util/getAnalyticsRequest.js'
 import { getDisplayPropertyUrl } from '../../../util/helpers.js'
 import { formatCount } from '../../../util/numbers.js'
 import EventPopup from './EventPopup.js'
